# backend/internal/api/v1/api.go
package v1

import (
	"encoding/json"
	"net/http"
	"time"

	"kingzecole/pay-by-bank-service/internal/api/server"
	"kingzecole/pay-by-bank-service/internal/models"
	"kingzecole/pay-by-bank-service/internal/service"
)

// API implements the ServerInterface generated by oapi-codegen
type API struct {
	accountService service.IAccountService
}

// NewAPI creates a new API instance
func NewAPI(accountService service.IAccountService) *API {
	return &API{
		accountService: accountService,
	}
}

// HealthCheck handles the health check endpoint
func (s *API) HealthCheck(w http.ResponseWriter, r *http.Request) {
	resp := server.HealthResponse{
		Status:    ptr("healthy"),
		Service:   ptr("pay-by-bank-service"),
		Version:   ptr("1.0.0"),
		Timestamp: ptr(time.Now().UTC().Format(time.RFC3339)),
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(resp)
}

// CreateAccount handles the account creation endpoint
func (s *API) CreateAccount(w http.ResponseWriter, r *http.Request) {
	var apiReq server.CreateAccountRequest
	if err := json.NewDecoder(r.Body).Decode(&apiReq); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(server.ErrorResponse{
			Success:   false,
			Message:   "Invalid request body",
			Timestamp: time.Now().UTC(),
		})
		return
	}

	// Map API request to Domain model
	domainReq := models.MandateRequest{
		User: models.User{
			FullName: apiReq.User.FullName,
			Email:    string(apiReq.User.Email),
			IBAN:     apiReq.User.Iban,
		},
		Bank: models.Bank{
			ID:   apiReq.Bank.Id,
			Name: apiReq.Bank.Name,
		},
	}
	if apiReq.Bank.Code != nil {
		domainReq.Bank.Code = *apiReq.Bank.Code
	}

	// Call Service
	result, err := s.accountService.CreateAccount(r.Context(), domainReq)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(server.ErrorResponse{
			Success:   false,
			Message:   err.Error(),
			Timestamp: time.Now().UTC(),
		})
		return
	}

	// Check for domain logic failure
	if !result.Success {
		w.WriteHeader(http.StatusBadRequest)
	} else {
		w.WriteHeader(http.StatusCreated)
	}

	// Map Domain result to API response
	resp := server.CreateAccountResponse{
		Success:   result.Success,
		Message:   result.Message,
		Timestamp: result.Timestamp,
		AccountId: &result.ID,
		Data: &map[string]interface{}{
			"accountHolder": result.User.FullName,
			"bankName":      result.Bank.Name,
		},
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(resp)
}

func ptr(s string) *string {
	return &s
}